Move code in `settings.js`
